requirements.txt
python3.6
anaconda3
pytorch=1.7
scikit-learn=0.19
torchtext=0.6
